Dedupe VerifyTransactionError into TransactionError
